Modified   ManaPHP/.phpstorm.meta.php Rename     ManaPHP/Caching/Cache.php (from ManaPHP/Cache.php) Rename     ManaPHP/Caching/Cache/Adapter/Apcu.php (from ManaPHP/Cache/Adapter/Apcu.php) Rename     ManaPHP/Caching/Cache/Adapter/Db.php (from ManaPHP/Cache/Adapter/Db.php) Rename     ManaPHP/Caching/Cache/Adapter/File.php (from ManaPHP/Cache/Adapter/File.php) Rename     ManaPHP/Caching/Cache/Adapter/Memory.php (from ManaPHP/Cache/Adapter/Memory.php) Rename     ManaPHP/Caching/Cache/Adapter/Redis.php (from ManaPHP/Cache/Adapter/Redis.php) Rename     ManaPHP/Caching/CacheInterface.php (from ManaPHP/CacheInterface.php) Rename     ManaPHP/Caching/MCache/ApcuCache.php (from ManaPHP/MCache/ApcuCache.php) Rename     ManaPHP/Caching/MCache/ArrCache.php (from ManaPHP/MCache/ArrCache.php) Rename     ManaPHP/Caching/MCacheInterface.php (from ManaPHP/MCacheInterface.php) Modified   ManaPHP/Cli/Controller.php Modified   ManaPHP/Component.php Modified   ManaPHP/Configuration/Settings/Adapter/Redis.php Modified   ManaPHP/Di.php Modified   ManaPHP/Di/FactoryDefault.php Modified   ManaPHP/Mvc/Controller.php Modified   ManaPHP/Mvc/Factory.php Modified   ManaPHP/Mvc/View/Widget.php Rename     tests/CachingCacheAdapterApcuTest.php (from tests/CacheAdapterApcuTest.php) Rename     tests/CachingCacheAdapterDbTest.php (from tests/CacheAdapterDbTest.php) Rename     tests/CachingCacheAdapterFileTest.php (from tests/CacheAdapterFileTest.php) Rename     tests/CachingCacheAdapterMemoryTest.php (from tests/CacheAdapterMemoryTest.php) Rename     tests/CachingCacheAdapterRedisTest.php (from tests/CacheAdapterRedisTest.php) Rename     tests/CachingCacheTest.php (from tests/CacheTest.php)

+++ tests/CachingCacheAdapterApcuTest.php
@@ -2,20 +2,22 @@
 
 namespace Tests;
 
-use ManaPHP\Cache\Adapter\Memory;
-use ManaPHP\Di\FactoryDefault;
+use ManaPHP\Caching\Cache\Adapter\Apcu;
 use PHPUnit\Framework\TestCase;
 
-class CacheAdapterMemoryTest extends TestCase
+class CachingCacheAdapterApcuTest extends TestCase
 {
-    public function setUp()
-    {
-        new FactoryDefault();
-    }
-
+    /**
+     * @requires  extension apc
+     */
     public function test_exists()
     {
-        $cache = new Memory();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
+
+        $cache = new Apcu();
         $cache->delete('var');
         $this->assertFalse($cache->exists('var'));
         $cache->set('var', 'value', 1000);
@@ -24,7 +26,12 @@ class CacheAdapterMemoryTest extends TestCase
 
     public function test_get()
     {
-        $cache = new Memory();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
+
+        $cache = new Apcu();
         $cache->delete('var');
 
         $this->assertFalse($cache->get('var'));
@@ -34,7 +41,12 @@ class CacheAdapterMemoryTest extends TestCase
 
     public function test_set()
     {
-        $cache = new Memory();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
+
+        $cache = new Apcu();
 
         $cache->set('var', '', 100);
         $this->assertSame('', $cache->get('var'));
@@ -49,12 +61,20 @@ class CacheAdapterMemoryTest extends TestCase
         $cache->set('var', 'value', 1);
         $this->assertTrue($cache->exists('var'));
         sleep(2);
-        $this->assertFalse($cache->exists('var'));
+        /**
+         * After the ttl has passed, the stored variable will be expunged from the cache (on the next request). If no ttl is supplied (or if the ttl is 0), the value will persist until it is removed from the cache manually, or otherwise fails to exist in the cache (clear, restart, etc.).
+         */
+        $this->assertTrue($cache->exists('var'));
     }
 
     public function test_delete()
     {
-        $cache = new Memory();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
+
+        $cache = new Apcu();
 
         //exists and delete
         $cache->set('var', 'value', 100);
